Reorganize protocols and add protocol attribute to modules
