--- client/protocols/__init__.py
@@ -1,8 +1,0 @@
-__all__ = [
-    'analytics',
-    'file_contents',
-    'grading',
-    'protocol',
-    'scoring',
-    'unlock',
-]
