+++ client/protocols/common/models.py
@@ -28,15 +28,3 @@ class Protocol(object):
     def on_interact(self):
         """Called to execute an interactive or output-intensive session."""
 
-def get_protocols(names):
-    mapping = {}
-    subclasses = Protocol.__subclasses__()
-    while subclasses:
-        protocol = subclasses.pop()
-        if protocol.name != Protocol.name:
-            mapping[protocol.name] = protocol
-        subclasses.extend(protocol.__subclasses__())
-    try:
-        return [mapping[name] for name in names]
-    except KeyError as e:
-        raise exceptions.OkException(str(e) + ' is not a protocol')
